requirements.txt
farm-haystack == 1.16
farm-haystack[ocr,pdf]==1.16.0
spacy==3.2.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz#egg=en_core_web_sm
matplotlib==3.5.1
nltk==3.7
numpy==1.22.1
pandas==1.4.0
pdfplumber==0.6.2
Pillow==9.1.1
seaborn==0.11.2
transformers==4.25.1
st-annotated-text==3.0.0
markdown==3.4.1
summa==1.2.0
plotly
xlsxwriter
altair==4.0
streamlit-aggrid
python-docx
setfit
plotly.express